splitting up ppis.js from annotations.js
